requirements.txt
elasticsearch==8.8.2
Flask==2.3.2
numpy==1.24.3
Pillow==10.0.0
Pyrebase==3.0.27
Pyrebase4==4.7.1
python-dotenv==1.0.0
tensorflow==2.13.0
tensorflow_macos==2.13.0
